pyproject.toml
[tool.poetry]
name = "api-fast"
version = "0.1.0"
description = ""
authors = ["Dante-Lucas <lucassilva270805@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
django-ninja = "^1.3.0"
poetry = "^1.8.3"
django-ninja-extra = {extras = ["swagger"], version = "^0.21.4"}
django-ninja-jwt = "^5.3.3"
python-decouple = "^3.8"
django-role-permissions = "^3.2.0"
uvicorn = "^0.30.6"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.13.0"

[tool.taskipy.tasks]
serve = "python manage.py runserver"
make = "python manage.py makemigrations"
migrate = "python manage.py migrate"
uvicorn = "uvicorn backend.asgi:application --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
